# Channel capacity