TODO:
* [x] Needs tests with all the possible capacities.
* [x] Benchmarks vs. the old implementation.

Change Channel's slot to be a slice

Allocating the Channel becomes a bit more complex as we need to manually
the layout for it and can't use the convenience of the Box type.

This also has the size effect of making the pointers in Sender, Receiver
and Manager a "fat" pointer. This means it will contain the regular
pointer and a length attribute, taking up (2 x 8) 16 bytes on 64bit.

This increases the size of Sender, Receiver, Manager from 8 to 16
bytes, and SendValue from 56 to 64 bytes.

If this size increase becomes a problem, or if we ever want to decrease
the memory usage we can change the fat pointer into a regular pointer
again and using some bits to indicate the length of the channel. For
examples if we have four different capacities, it can be encoded into
just 2 bits. These two bits can ORed with the regular pointer and be
received later.

The upside of such an approach would be that the size is reduced from 16
to 8 bytes. The downside is that on each access to the channel the
reference has to be prepared by unpacking the length from the pointer.
And ensuring the length is correct at all times is an additional
maintenance burden.


Add `new` function

Allows the creation of an user specified capacity bounded channel. The
current implementation is limited to a capacity in the range 1..=29.

This limitation is because of the size of the `status` in the Channel.
As each slot needs two bits for its status and we need enough bits for
the receiver position the current limit is 29. The following must hold
true, where $n is the capacity of the channel:
2 ^ (64 - ($N * 2)) >= $N.

Changes channel status to an `AtomicU64`.

Closes #4.