@Thomasdezeeuw 2021-02-22
This still needs test with all the capacities.