@Thomasdezeeuw 2021-03-01
The performance is much worse, a good 20%, from ~180ns to ~230ns.

This pr:

```
send/usize/inbox        time:   [230.38 ns 230.89 ns 231.45 ns]
                        change: [-1.0901% -0.6502% -0.1808%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 14 outliers among 100 measurements (14.00%)
  10 (10.00%) high mild
  4 (4.00%) high severe

send/unit/inbox         time:   [233.16 ns 233.69 ns 234.24 ns]
                        change: [-0.0616% +0.4502% +0.9590%] (p = 0.08 > 0.05)
                        No change in performance detected.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe
```

Current master (i.e. single capacity), relative to this pr:

```
send/usize/inbox        time:   [178.58 ns 178.86 ns 179.13 ns]
                        change: [-22.828% -22.495% -22.137%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  5 (5.00%) high severe

send/unit/inbox         time:   [179.05 ns 179.73 ns 180.62 ns]
                        change: [-23.513% -23.135% -22.783%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  3 (3.00%) high mild
  10 (10.00%) high severe
```