@Thomasdezeeuw 2021-08-14
Continued in https://github.com/Thomasdezeeuw/inbox/pull/30.